+++ fortnite_wiki/fortnite_wiki/Helpers/WeaponsLevels.swift
@@ -20,7 +20,7 @@ class WeaponsLevels: NSObject {
     }
 }
 
-class UIBackgroundView: NSObject {
+class FormatLevels: NSObject {
 
     let shadows = ShadowLayers()
     let color0 = [HexColor("969696")!, HexColor("4A4A4A")!]
@@ -29,7 +29,7 @@ class UIBackgroundView: NSObject {
     let color3 = [HexColor("D257FF")!, HexColor("530080")!]
     let color4 = [HexColor("FFD528")!, HexColor("805600")!]
 
-    func formatUIBackgroundViewFromLevel(view : UIView, level: Int) {
+    func formatUIBackgroundViewFromLevel(view: UIView, level: Int) {
         
         switch level {
         case 0:
@@ -55,6 +55,47 @@ class UIBackgroundView: NSObject {
         view.layer.borderWidth = 8.5
         view.layer.borderColor = UIColor.flatWhite.cgColor
         shadows.setLayer(view: view)
+
+        view.layer.cornerRadius = 6.5
+    }
+
+    func formatCellGradients(cell: UICollectionViewCell, levels: [Int]) {
+        switch levels {
+        case [0,1,2]:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [UIColor.init(hexString: "969696")!, UIColor.init(hexString: "969696")!, UIColor.init(hexString: "4FCA00")!, UIColor.init(hexString: "00BFFF")!])
+            break
+        case [1,2]:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [UIColor.init(hexString: "4FCA00")!, UIColor.init(hexString: "00BFFF")!])
+            break
+        case [3,4]:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [UIColor.init(hexString: "B83DF2")!, UIColor.init(hexString: "E6BB0E")!])
+        default:
+            cell.backgroundColor = UIColor.black
+            break
+        }
+    }
+
+    func formatCellGradient(cell: UICollectionViewCell, level: Int) {
+        switch level {
+        case 0:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("4A4A4A")!, HexColor("969696")!])
+            break
+        case 1:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("037E00")!, HexColor("69E41A")!])
+            break
+        case 2:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("004080")!, HexColor("00BFFF")!])
+            break
+        case 3:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("530080")!, HexColor("D257FF")!])
+            break
+        case 4:
+            cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("805600")!, HexColor("FFD528")!])
+            break
+        default:
+            cell.backgroundColor = UIColor.black
+            break
+        }
     }
     
     override var description: String {